+++ src/structures/BuilderBase.cpp
@@ -36,6 +36,7 @@
 #include <algorithm>
 
 const int BuilderBase::itemOrder[] = {    Structure_Slab4, Structure_Slab1, Structure_Road, Structure_IX, Structure_StarPort,
+                                           Structure_LoveFactory,
                                            Structure_HighTechFactory, Structure_HeavyFactory, Structure_RocketTurret,
                                            Structure_Scoutpost,
                                            Structure_RepairYard, Structure_GunTurret, Structure_TechCenter, Structure_WOR,